requirements-cpu.txt
# io libraries 
rich
click
loguru

# vectors processing libraries
numpy
pandas
networkx 
opencv-python

# socket rest-api libraries 
pyzmq
jinja2
fastapi
uvicorn
websockets
python-multipart

# graph vision deep learning libraries 
spacy 
torch 
torchtext 
torchvision
sentence-transformers 
